refactor: Move the unassign app into its own isolated express app
This change follows the pattern we've started to use for the "new" move
app and the person escort record/youth risk assessment.

It isolates the logic around unassigning a move. This should help with
maintaining the move application as it has grown considerably.
