+++ app/move/config/index.js
@@ -1,9 +1,7 @@
 const assignConfig = require('./assign.config')
-const unassignConfig = require('./unassign.config')
 const updateConfig = require('./update.config')
 
 module.exports = {
   assignConfig,
-  unassignConfig,
   updateConfig,
 }
