+++ app/move/controllers/index.js
@@ -1,7 +1,6 @@
 const assign = require('./assign')
 const confirmation = require('./confirmation')
 const timeline = require('./timeline')
-const Unassign = require('./unassign')
 const update = require('./update')
 const view = require('./view')
 
@@ -9,7 +8,6 @@ module.exports = {
   assign,
   confirmation,
   timeline,
-  Unassign,
   update,
   view,
 }
