+++ app/move/steps/index.js
@@ -1,9 +1,7 @@
 const assign = require('./assign')
-const unassign = require('./unassign')
 const update = require('./update')
 
 module.exports = {
   assign,
-  unassign,
   update,
 }
